+++ CHANGELOG.md
@@ -39,6 +39,7 @@ but may switch to [CalVer Versioning](https://calver.org/) in the future.
 #### Changed
 
 - Made block comments delimited by `#*` and `*#` instead of `##`, allowing for nested block comments
+- Made line comments start with `##` instead of `#`
 - Made `;` the array's items separator
 - Made `!=` -> `!==` (boolean comparison "not equals" operator is now "not equals equals" operator)